Move `CalendarFactoryInterface` and `CalendarFactory` to `Calendar` namespace
